requirements.txt
requests==2.28.2
pydantic==1.10.7
python-dotenv==0.20.0
httpx==0.23.0
tenacity==8.0.1
rich==12.6.0
click==7.1.2
orjson==3.6.8
typing-extensions==4.3.0
pytest==7.1.2
